Changed UserController from admin (/Admin) and from user profile.
