+++ app/Http/Controllers/UserController.php
@@ -2,96 +2,63 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
-
 use App\User;
+use App\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 
 class UserController extends Controller {
 	/**
-	 * Display a listing of the resource.
+	 * Show the form for editing the current user
 	 *
-	 * @return \Illuminate\View\View
+	 * @return \Illuminate\Http\Response
 	 */
-	public function index(Request $request) {
-		$keyword = $request->get('search');
-		$perPage = 25;
-
-		if (!empty($keyword)) {
-			$users = User::latest()->paginate($perPage);
-		} else {
-			$users = User::latest()->paginate($perPage);
-		}
-
-		return view('test.users.index', compact('users'));
-	}
-
-	/**
-	 * Show the form for creating a new resource.
-	 *
-	 * @return \Illuminate\View\View
-	 */
-	public function create() {
-		return view('test.users.create');
-	}
-
-	/**
-	 * Store a newly created resource in storage.
-	 *
-	 * @param \Illuminate\Http\Request $request
-	 *
-	 * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
-	 */
-	public function store(Request $request) {
-
-		$requestData = $request->all();
-
-		User::create($requestData);
-
-		return redirect('users')->with('flash_message', 'User added!');
-	}
-
-	/**
-	 * Display the specified resource.
-	 *
-	 * @param  int  $id
-	 *
-	 * @return \Illuminate\View\View
-	 */
-	public function show($id) {
-		$user = User::findOrFail($id);
-
-		return view('test.users.show', compact('user'));
-	}
-
-	/**
-	 * Show the form for editing the specified resource.
-	 *
-	 * @param  int  $id
-	 *
-	 * @return \Illuminate\View\View
-	 */
-	public function edit($id) {
-		$user = User::findOrFail($id);
-
-		return view('test.users.edit', compact('user'));
+	public function edit() {
+		return view('user.edit', [
+			'skills' => Tag::where('type', 1)->get(),
+			'user' => Auth::user()
+		]);
 	}
 
 	/**
 	 * Update the specified resource in storage.
 	 *
-	 * @param \Illuminate\Http\Request $request
-	 * @param  int  $id
-	 *
-	 * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+	 * @param  \Illuminate\Http\Request  $request
+	 * @param  \App\User  $user
+	 * @return \Illuminate\Http\Response
 	 */
-	public function update(Request $request, $id) {
+	public function update(Request $request) {
+		$attributes = request()->validate([
+			'name' => ['present']
+		]);
 
-		$requestData = $request->all();
+		$user = Auth::user();
+		$user->name = $attributes['name'];
+		$user->save();
 
-		$user = User::findOrFail($id);
-		$user->update($requestData);
+		// Update the picture of the user
+		if (isset($request->avatar)) {
+			$this->updateImg($request->file('avatar'));
+		}
 
-		return redirect('users')->with('flash_message', 'User updated!');
+		// Update skillset of the user
+		$user->skillset()->detach();
+		if (isset($request->skills)) {
+			$user->skillset()->attach($request->skills);
+		}
+
+		return redirect()->back()->with([
+			'skills' => Tag::where('type', 1)->get(),
+			'user' => Auth::user()
+		]);
+	}
+
+	private function updateImg($image) {
+		$path = isset($image) ? Storage::putFile('public', $image) : Auth::user()->picture_url;
+		$user = Auth::user();
+		$user->picture_url = Storage::url($path);
+		$user->save();
 	}
 }
